vendor filtering by admin and and user
Added a filter in the vendor list table:

If the user logged in is an admin he is able to see all vendors.

If the user logged in is a regular user then he can only see his vendors.

Notifications fix was removed due to duplicate messages bug a rollback has been made.